feat: (#5) 노션 상세페이지로 이동하는 훅, 바텀시트 관련 상태 훅 분리, 리팩토링
