+++ app/notion/[id]/page.tsx
@@ -1,39 +1,33 @@
 'use client';
 
-import React, { ReactNode, useState } from 'react';
+import { useRecentlyNotionContext } from '@components/context/RecentlyNotionContext';
 
-import BottomSheet from '@components/common/BottomSheet';
 import CircleLine from '@components/common/CircleLine';
 import Description from '@components/common/Description';
 import Title from '@components/common/Title';
 import NotionList from '@components/item/NotionList';
-import NotionForm from '@components/item/NotionForm';
 
 import { mockNotionBanana } from '@mocks/mockData/notionList';
 
-import { useRouter } from 'next/navigation';
-import { useRecentlyNotionContext } from '@components/context/RecentlyNotionContext';
-
-type Content = 'notionItemForm' | 'notionItemPlusMenu';
-
-const content: Record<Content, ReactNode> = {
-  notionItemForm: <NotionForm />,
-  notionItemPlusMenu: <></>,
-};
+import { useBottomSheetContent } from 'hooks/useBottomSheetContent';
+import { useMovePage } from 'hooks/useMovePage';
 
 export default function play({ params }: { params: { id: number } }) {
-  const router = useRouter();
   const { addNotionItem } = useRecentlyNotionContext();
-  const [bottomSheetContent, setBottomSheetContent] = useState<Content | null>(
-    null,
-  );
+  const { moveNotionItemPage } = useMovePage();
+
+  const {
+    handlePlusButtonClick,
+    handleMoreMenuButtonClick,
+    bottomSheetComponent,
+  } = useBottomSheetContent();
 
   //추후 fetch한 데이터로 수정
   const { id, name, description, relatedNotionList } = mockNotionBanana;
 
-  const moveNotionItemPage = (id: number) => {
+  const handleNotionItemClick = () => (id: number) => {
     addNotionItem({ id, name });
-    router.push(`/notion/${id}`, { scroll: true });
+    moveNotionItemPage(id);
   };
 
   return (
@@ -43,17 +37,11 @@ export default function play({ params }: { params: { id: number } }) {
       <Description content={description} />
       <NotionList
         notionList={relatedNotionList}
-        handlePlusButtonClick={() => setBottomSheetContent('notionItemForm')}
-        handleMoreMenuButtonClick={() =>
-          setBottomSheetContent('notionItemPlusMenu')
-        }
-        handleNotionItemClick={moveNotionItemPage}
+        handlePlusButtonClick={handlePlusButtonClick}
+        handleMoreMenuButtonClick={handleMoreMenuButtonClick}
+        handleNotionItemClick={() => handleNotionItemClick()}
       />
-      {bottomSheetContent && (
-        <BottomSheet closeEvent={() => setBottomSheetContent(null)}>
-          {content[bottomSheetContent]}
-        </BottomSheet>
-      )}
+      {bottomSheetComponent}
     </main>
   );
 }
